requirements/ci.txt
django-filter
django-guardian
djangorestframework
djangorestframework-guardian
unittest-xml-reporting
pytest
pytz
coverage
